Refactor Google Maps embed logic
Utilize the Google Maps Embed API so that we can dynamically embed the
map based on the place rather than have to generate the embed URL for
each venue and put it into the frontmatter for the event.

Also maps link to take you to the Google Maps page for the current event
address (was hardcoded to Code for America).